python/requirements.txt
futures;python_version<="2.7"
python-dateutil
requests>=2.9.2
vcrpy<2.0.0 #2.0 breaks python 3.4
azure-common
cryptography>=2.3
adal
html5lib
lxml
asyncio
applicationinsights
azure-cognitiveservices-vision-computervision>=0.3.0  # sample won't work with previous versions
pandas
google-cloud-vision==1.0.0